+++ cpm/generators/__init__.py
@@ -2,10 +2,10 @@
 Here we import all the modules that are needed for the project.
 """
 
-__all__ = ["Wrapper", "Simulator", "Parameters", "Value"]
+__all__ = ["Wrapper", "Simulator", "Parameters", "Value", "LogParameters"]
 
 from .simulator import Simulator
 from .wrapper import Wrapper
-from .parameters import Parameters, Value
+from .parameters import Parameters, Value, LogParameters
 
 del simulator, wrapper, parameters