+++ cpm/optimisation/__init__.py
@@ -1,7 +1,10 @@
-__all__ = ["DifferentialEvolution", "Fmin", "Bads", "minimise"]
+
+__all__ = ["DifferentialEvolution", "Fmin", "FminBound", "Bads", "minimise"]
 
 
 from .genetic import DifferentialEvolution
-from .fmin import Fmin
+from .fmin import Fmin, FminBound
+from .free import Minimize
+from .empirical import EmpiricalBayes
 from .bads import Bads
 from .minimise import *