+++ cpm/optimisation/genetic.py
@@ -2,6 +2,8 @@ from scipy.optimize import differential_evolution
 import pandas as pd
 import numpy as np
 import copy
+import warnings
+
 from . import minimise
 from . import utils
 from ..generators import Simulator, Wrapper
@@ -69,14 +71,17 @@ class DifferentialEvolution:
         self.parameters = []
         self.participant = data[0]
         if isinstance(model, Wrapper):
-            self.parameter_names = self.function.parameter_names
+            self.parameter_names = self.function.parameters.free()
         if isinstance(model, Simulator):
-            self.parameter_names = self.function.function.parameter_names
-        if hasattr(self.function, "bounds"):
-            self.bounds = self.function.bounds
-        else:
-            self.bounds = bounds
-            # raise ValueError("You must define the parameter bounds in the Model object.")
+            raise ValueError(
+                "The DifferentialEvolution algorithm is not compatible with the Simulator object."
+            )
+        self.bounds = bounds
+        if bounds is None:
+            self.bounds = self.function.parameter.bounds()
+            warnings.warn(
+                "No explicit bounds are specified. Using model's parameter specification."
+            )
 
     def minimise(self, pars, **args):
         """